requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
decorator==4.4.2
Flask==1.1.2
idna==2.9
imageio==2.8.0
imageio-ffmpeg==0.4.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
moviepy==1.0.0
numpy==1.18.2
Pillow==7.1.1
pkg-resources==0.0.0
proglog==0.1.9
pytube3==9.6.4
requests==2.23.0
tqdm==4.45.0
typing-extensions==3.7.4.2
urllib3==1.25.9
Werkzeug==1.0.1
WTForms==2.2.1
